background_extraction/bg_requirements.txt
imageio==2.8.0
imageio-ffmpeg==0.4.1
numpy==1.18.2
Pillow==8.1.1